API made core and above, role changed from string to enum, moved EmailRequest to model, Tested api
