+++ src/main/java/com/pecacm/backend/model/EmailRequest.java
@@ -1,16 +1,21 @@
-package com.pecacm.backend.request;
+package com.pecacm.backend.model;
 
+import com.pecacm.backend.enums.Role;
 import jakarta.annotation.Nullable;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.util.List;
 
 @Getter
 @Setter
+@AllArgsConstructor
+@NoArgsConstructor
 public class EmailRequest {
     @Nullable
-    private String role;
+    private Role role;
     @Nullable
     private List<String> emails;
     private String subject;
